requirements.txt
absl-py==2.3.1
astunparse==1.6.3
beautifulsoup4==4.14.2
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
cursor==1.3.5
cycler==0.12.1
deepface==0.0.95
dlib==20.0.0
face-alignment==1.4.1
face-recognition==1.3.0
face-recognition-models==0.3.0
facenet-pytorch==2.6.0
filelock==3.19.1
fire==0.7.1
flask==3.1.2
flask-cors==6.0.1
flatbuffers==25.9.23
fonttools==4.60.1
fsspec==2025.9.0
gast==0.6.0
gdown==5.2.0
google-pasta==0.2.0
grpcio==1.76.0
gunicorn==23.0.0
h5py==3.15.1
idna==3.11
imageio==2.37.2
itsdangerous==2.2.0
jinja2==3.1.6
joblib==1.5.2
keras==3.12.0
kiwisolver==1.4.9
lazy-loader==0.4
libclang==18.1.1
llvmlite==0.45.1
lz4==4.4.5
markdown==3.10
markdown-it-py==4.0.0
markupsafe==3.0.3
matplotlib==3.10.7
mdurl==0.1.2
ml-dtypes==0.5.4
mpmath==1.3.0
mtcnn==1.0.0
namex==0.1.0
networkx==3.5
numba==0.62.1
numpy==1.26.4
nvidia-cublas==13.0.0.19
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti==13.0.48
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvcc-cu12==12.9.86
nvidia-cuda-nvrtc==13.0.48
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime==13.0.48
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cudnn-cu13==9.13.0.50
nvidia-cufft==12.0.0.15
nvidia-cufft-cu12==11.0.2.54
nvidia-cufile==1.15.0.42
nvidia-curand==10.4.0.35
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver==12.0.3.29
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse==12.6.2.49
nvidia-cusparse-cu12==12.1.0.106
nvidia-cusparselt-cu13==0.8.0
nvidia-nccl-cu12==2.19.3
nvidia-nccl-cu13==2.27.7
nvidia-nvjitlink==13.0.39
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvshmem-cu13==3.3.24
nvidia-nvtx==13.0.39
nvidia-nvtx-cu12==12.1.105
opencv-python==4.11.0.86
opt-einsum==3.4.0
optree==0.18.0
packaging==25.0
pandas==2.3.3
pillow==10.2.0
protobuf==6.33.1
psycopg2-binary==2.9.11
pygments==2.19.2
pyparsing==3.2.5
pysocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
retina-face==0.0.17
rich==14.2.0
scikit-image==0.25.2
scipy==1.16.3
setuptools==80.9.0
six==1.17.0
soupsieve==2.8
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.2.0
tf-keras==2.20.1
tifffile==2025.10.16
torch==2.2.2
torchaudio==2.9.1+cu130
torchvision==0.17.2
tqdm==4.67.1
triton==3.5.1
typing-extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
werkzeug==3.1.3
wheel==0.45.1
wrapt==2.0.1
